requirements.txt
faster-whisper
sounddevice
soundfile
numpy
scipy
python-dotenv
customtkinter
pynput
